Refactor some of the emit graph into a new module
Yay for TypeFamilies for allowing us to do some absolutely absurd things
here.

This effectively abstracts out the whole graph handling of the backend
into a separate module. This should allow us to reuse it for working
with any other backends which should generate a simple syntax tree (C,
JS, Scheme, etc...)
